Remove piano_dialogue_server
