+++ AGENTS.md
@@ -10,7 +10,7 @@
 - visionOS App：`LonelyPianistAVP/`
 - visionOS 测试：`LonelyPianistAVPTests/`（Swift Testing）
 - SwiftPM 包：`Packages/RealityKitContent/`
-- Python 后端：`piano_dialogue_server/`
+- Python 后端：`piano_duet_server/`
 - 规划/知识库：`.github/features/`、`docs/`
 
 # Apple App 开发规范 for AI（Swift/SwiftUI 基线）
