+++ docs/configuration.md
@@ -7,7 +7,6 @@
 | Xcode 工程 | `LonelyPianist.xcodeproj` | 包含 macOS app、visionOS app 与测试 target。 |
 | macOS scheme | `LonelyPianist` | recorder app 与 `LonelyPianistTests`。 |
 | visionOS scheme | `LonelyPianistAVP` | AVP app 与 `LonelyPianistAVPTests`。 |
-| Python 服务 | `piano_dialogue_server/scripts/run_server.sh` | 创建 `.venv`、安装依赖并启动 uvicorn。 |
 | Python 服务（Duet） | `piano_duet_server/scripts/run_server.sh` | 创建 `.venv`、安装依赖并启动 uvicorn（可选 Magenta）。 |
 
 当前仓库没有 `.github/workflows/`，自动化验证以本地命令为准。
@@ -27,7 +26,7 @@
 | 配置面 | 位置 | 说明 |
 | --- | --- | --- |
 | 权限说明 | `LonelyPianistAVP/Resources/Info.plist` | Hand Tracking、World Sensing、Microphone、Bluetooth、Local Network。 |
-| Bonjour | `NSBonjourServices` | `_lonelypianist._tcp`（Dialogue）与 `_lpduet._tcp`（Duet），用于发现本地 Python 后端。 |
+| Bonjour | `NSBonjourServices` | `_lpduet._tcp`（Duet），用于发现本地 Python 后端。 |
 | ATS local networking | `NSAppTransportSecurity` | 允许局域网 HTTP 连接。 |
 | MusicXML 文件类型 | `UTImportedTypeDeclarations` | 导入 `.musicxml` / `.xml`。 |
 | 字体 | `UIAppFonts` | `Bravura.otf`。 |
@@ -50,20 +49,7 @@
 - “练习判定：左右手分别满足”已改为强制启用（不再作为可配置项，也不再暴露 UI 开关）。
 - 当前仍会通过 `UserDefaults` 保存练习手（左/右/双手）、手动推进方式与音频识别模式等设置项。
 
-## Python 服务配置
-
-| 项 | 默认值 / 位置 | 说明 |
-| --- | --- | --- |
-| host | `0.0.0.0` in `piano_dialogue_server/scripts/run_server.sh` | 便于 AVP 真机访问。 |
-| port | `8765` | HTTP、WebSocket 与 Bonjour 广播使用同一端口。 |
-| Bonjour service type | `_lonelypianist._tcp.local.` | `piano_dialogue_server/server/media/bonjour.py` 与 AVP discovery 对齐。 |
-| `AMT_MODEL_DIR` | 无默认硬编码目录优先级之一 | 本地模型目录。 |
-| `AMT_MODEL_ID` | `stanford-crfm/music-large-800k` | Hugging Face 模型 ID。 |
-| `AMT_DEVICE` | 自动选择 | `mps` / `cuda` / `cpu`。 |
-| `HF_ENDPOINT` | 可选 | Hugging Face 镜像地址。 |
-| `DIALOGUE_DEBUG` | unset | 设为 `1` 时写调试包。 |
-
-### Duet server（piano_duet_server）
+## Python 服务配置（Duet）
 
 | 项 | 默认值 / 位置 | 说明 |
 | --- | --- | --- |
