+++ docs/data-flow.md
@@ -1,6 +1,6 @@
 # 数据流
 
-本文只描述当前代码存在的运行链路。macOS 端不再包含 MIDI mapping、键盘注入或 Dialogue client；visionOS 端（AVP）的 AI 即兴链路支持 **本地生成后端**（Swift）与 **可选网络后端**（本机 Python 服务），并且严格只使用用户在 practice 设置中选择的后端（不会自动降级/切换）。
+本文只描述当前代码存在的运行链路。macOS 端不再包含 MIDI mapping、键盘注入或 AVP 的网络后端 client；visionOS 端（AVP）的 AI 即兴链路支持 **本地生成后端**（Swift）与 **可选网络后端**（本机 Python 服务），并且严格只使用用户在 practice 设置中选择的后端（不会自动降级/切换）。
 
 ## 主流程
 
@@ -14,7 +14,7 @@
 | AVP 曲库 | bundled MusicXML / 用户导入 MusicXML | `SongLibraryViewModel` + `PracticePreparationService` | `PreparedPractice` |
 | AVP 练习 | `PreparedPractice` + selected piano mode | `ARGuideViewModel` + `PracticeSessionViewModel` | 步骤推进、谱面、高亮、录制与回放 |
 | AVP AI 即兴 | recorded phrase / selected clip | `AIPerformanceService` + `ImprovBackendRegistry` | 生成片段并排程回放（严格按所选后端） |
-| Python 生成（可选） | HTTP / WebSocket request | FastAPI + model engine | 仅当选择 `网络本地连接` 后端时触发 |
+| Python 生成（可选） | HTTP request | FastAPI + engine | 仅当选择 `网络本地连接（A.I. Duet）` 后端时触发 |
 
 ## macOS recorder
 
@@ -100,7 +100,7 @@ flowchart TD
 
 practice 窗口的 settings popover 中可选择后端：
 
-- `网络本地连接`：通过 Bonjour 发现 + HTTP 请求调用本机 `piano_dialogue_server`（电脑端运行）。
+- `网络本地连接（A.I. Duet）`：通过 Bonjour 发现 + HTTP 请求调用本机 `piano_duet_server`（电脑端运行）。
 - `本地规则生成（Local rule）`：AVP 端直接调用 SwiftPM `ImprovEngines`（seed 可复现）。
 - `按谱片段回放（tick-range replay）`：不做生成，回放当前谱面片段；它不是自动 fallback，只会在用户选择时使用。
 
@@ -111,7 +111,7 @@ sequenceDiagram
   participant Backend as ImprovBackendProtocol
   participant Engines as ImprovEngines (SwiftPM)
   participant Bonjour as BonjourBackendDiscoveryService
-  participant API as piano_dialogue_server
+  participant API as piano_duet_server
 
   Settings-->>AVP: selected ImprovBackendKind
   AVP->>Backend: generatePlaybackPlan(request)
@@ -120,7 +120,7 @@ sequenceDiagram
     Engines-->>Backend: generated notes
     Backend-->>AVP: schedule
   else network bonjour http
-    Backend->>Bonjour: discover _lonelypianist._tcp.local.
+    Backend->>Bonjour: discover _lpduet._tcp.local.
     Bonjour-->>Backend: host, port, path=/generate
     Backend->>API: POST /generate
     API-->>Backend: ResultResponse
@@ -130,4 +130,4 @@ sequenceDiagram
   end
 ```
 
-Python 后端的 `GenerateParams.strategy` 仅支持 `model`；本地规则生成已迁移到 SwiftPM（`Packages/ImprovEngines/`），仅用于 AVP 端的本地生成路径。
+Python 后端是可选网络后端；本地规则生成已迁移到 SwiftPM（`Packages/ImprovEngines/`），仅用于 AVP 端的本地生成路径。
